original project current state
Hey, decided to skip the redo, so here is my full Unity/C# 3rd person practice project in its current state.

Implemented:
-basic movement controls
-jumping
-shooting projectiles
-telekinesis
-levitation
-dashing.

Those are all set up and called in the PlayerCharacter script/class, on isKeyDown.

Implemented:
-basic UI (health/energy, score, info text)
-a basic camera boom
-pause & options menu
-audio volume controls
-mouse sensitivity controls
-photo capture/very basic photo mode

Implemented:
-saving/loading user data (screenshots path) with JsonUtility.FromJson/.ToJson

This is done via the MainManager script/class, which uses DontDestroyOnLoad to persist during level loads and sort of acts as blackboard for the other scripts/passes encapsulated public variable values/states between the scripts.

Implemented:
-basic movement controls for a flying vehicle
-switching between player and vehicle POV on isKeyDown
-additional "hoverboard" vehicle on isKeyDown
-switching textures on the  "hoverboard" vehicle on isKeyDown

Implemented:
-jump pads
-accelerators

There are also some "enemies"/hazards, which move randomly and "attack" when in close proximity to the player.

And lastly ...

implemented:
-a basic "laser beam" which reflects off objects tagged as "Mirror"

That´s pretty much it.

Already found & fixed a bagfull of bugs and optimized the code a bit, but there is still a lot to clean up/fix and I´ll probably add more features, a decent level and some sort of a game concept soon.